change price field in Product class from 'PositiveIntegerField' to 'DecimalField'

+++ mysite/goods/models.py
@@ -3,7 +3,7 @@ from django.db import models
 
 class Product(models.Model):
     name = models.CharField(max_length=128, verbose_name='Название товара')
-    price = models.PositiveIntegerField(verbose_name='Цена товара')
+    price = models.DecimalField(max_digits=10, decimal_places=2, verbose_name='Цена товара')
 
     def __str__(self):
         return f'{self.name} - {self.price} руб.'
